refactor spec helpers
