+++ card/spec/card_spec_helper.rb
@@ -1,8 +1,31 @@
 class Card
+  def self.gimme! name, args={}
+    Card::Auth.as_bot do
+      c = Card.fetch(name, new: args)
+      c.putty args
+      Card.fetch name
+    end
+  end
+
+  def self.gimme name, args={}
+    Card::Auth.as_bot do
+      c = Card.fetch(name, new: args)
+      if args[:content] && c.content != args[:content]
+        c.putty args
+        c = Card.fetch name
+      end
+      c
+    end
+  end
+
+  cattr_accessor :rspec_binding
+
   module SpecHelper
     include Rails::Dom::Testing::Assertions::SelectorAssertions
+
     # ~~~~~~~~~  HELPER METHODS ~~~~~~~~~~~~~~~#
 
+    include Card::Model::SaveHelper
     def login_as user
       Card::Auth.current_id = (uc = Card[user.to_s]) && uc.id
       return unless @request
@@ -13,6 +36,24 @@ class Card
 
     def create! name, content=""
       Card.create! name: name, content: content
+    end
+
+    def create_or_update name_or_args, args={}
+      Card::Auth.as_bot { super }
+    end
+
+    def update name, args
+      Card::Auth.as_bot { update_card name, args }
+    end
+
+    def putty args={}
+      Card::Auth.as_bot do
+        if args.present?
+          update_attributes! args
+        else
+          save!
+        end
+      end
     end
 
     def assert_view_select view_html, *args, &block
@@ -132,6 +173,43 @@ class Card
         remove_test_event stage, name
       end
       Card.rspec_binding = false
+    end
+
+    def bucket_credentials key
+      @buckets ||= begin
+        yml_file =
+          ENV["BUCKET_CREDENTIALS_PATH"] ||
+            File.expand_path("../config/bucket_credentials.yml", __FILE__)
+        File.exist?(yml_file) ? YAML.load_file(yml_file).deep_symbolize_keys : {}
+      end
+      @buckets[key]
+    end
+
+    # rubocop:disable Lint/Eval
+    def method_missing m, *args, &block
+      return super unless Card.rspec_binding
+      suppress_name_error do
+        method = eval("method(%s)" % m.inspect, Card.rspec_binding)
+        return method.call(*args, &block)
+      end
+      suppress_name_error do
+        return eval(m.to_s, Card.rspec_binding)
+      end
+      super
+    end
+    # rubocop:enable Lint/Eval
+
+    def suppress_name_error
+      yield
+    rescue NameError
+    end
+
+    def format_with_set set, format_type=:html
+      singleton_class.send :include, set
+      format = format format_type
+      format_class = Card::Format.format_class_name format_type
+      format.singleton_class.send :include, set.const_get(format_class)
+      yield(format)
     end
   end
 end
